Allow font ligatures to work on Windows
This change allows fonts like Fira Code, that support font ligatures, to
work on Windows.

How does this work?

* I've added a true/false int value to the settings.
* If it is true, and a font supports ligatures, they will be drawn.
* Instead of calling `ExtTextOut` directly, we call the new `textout_with_ligatures` function.
* `textout_with_ligatures` calls `GetCharacterPlacement` with `GCP_LIGATE` first, then calls  `ExtTextOut`.
* `GetCharacterPlacement` will fill up a new buffer with the font characters after ligatures have been applied.

Note that I've added the new setting in the Window/Translation pane. It
might be better in the Window/Appearance pane under the Font Settings
section. However, I can't determine how to get the dialog box to be
larger to accomodate the additional row.